Expose AudioDriverDummy speaker_mode in settings.
When the Dummy driver is selected, it will register the property
'audio/driver/speaker_mode', which is read during init, if preset.
It defaults to stereo. Because this is in the dummy driver `init()`,
this setting doesn't appear in the editor if any other driver is
active.

Also changed get_channels to give the correct number of channels for
5.1 and 7.1.