Add support for audio channel remapping.
The AudioServer processes audio samples as a number of stereo channels,
which are copied to the audio driver buffers. This commit adds
the ability to remap AudioServer channels to different driver
channels, primarily for debugging purposes. The way it is configured
is with get/set methods on AudioServer, exposed in GDScript/C#.

This depends on the speaker_mode override PR #105934.

It only allows for remapping the stereo channels Godot works with, and
not for mixing left/right samples between channels.